Rename ScopeUtil to LifecycleScopes
