+++ autodispose/src/main/java/com/uber/autodispose/LifecycleScopes.java
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2017. Uber Technologies
+ * Copyright (C) 2018. Uber Technologies
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -25,25 +25,25 @@ import io.reactivex.functions.Predicate;
 import java.util.concurrent.Callable;
 
 /**
- * Utilities for dealing with scopes, usually for providers. This includes factories for resolving
+ * Utilities for dealing with {@link LifecycleScopeProvider}s. This includes factories for resolving
  * {@link Maybe} representations of scopes, corresponding events, etc.
  */
-public final class ScopeUtil {
+public final class LifecycleScopes {
 
   private static final Function<Object, LifecycleEndNotification> TRANSFORM_TO_END =
       new Function<Object, LifecycleEndNotification>() {
-        @Override public LifecycleEndNotification apply(Object o) throws Exception {
+        @Override public LifecycleEndNotification apply(Object o) {
           return LifecycleEndNotification.INSTANCE;
         }
       };
 
   private static final Predicate<Boolean> IDENTITY_BOOLEAN_PREDICATE = new Predicate<Boolean>() {
-    @Override public boolean test(Boolean b) throws Exception {
+    @Override public boolean test(Boolean b) {
       return b;
     }
   };
 
-  private ScopeUtil() {
+  private LifecycleScopes() {
     throw new InstantiationError();
   }
